requirements.txt
alembic==1.6.5
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==8.0.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.22
Pillow==8.2.0
protobuf==3.17.1
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.17
SQLAlchemy-Utils==0.36.8
Werkzeug==2.0.1
wtf==0.1
WTForms==2.3.3
